ch4/generic: move generic comm function to separate file
Creating file for all MPIDIG_comm_* functions.
